+++ packages/ice/src/config.ts
@@ -61,7 +61,7 @@ const userConfig = [
   },
   {
     name: 'dataLoader',
-    validation: 'boolean',
+    validation: 'boolean|object',
     defaultValue: true,
   },
   {