+++ packages/ice/src/createService.ts
@@ -108,6 +108,8 @@ async function createService({ rootDir, command, commandArgs }: CreateServiceOpt
       dataCache,
     },
   });
+  // Load .env before resolve user config, so we can access env variables defined in .env files.
+  await setEnv(rootDir, commandArgs);
   // resolve userConfig from ice.config.ts before registerConfig
   await ctx.resolveUserConfig();
 
@@ -140,7 +142,6 @@ async function createService({ rootDir, command, commandArgs }: CreateServiceOpt
   const { routes: routesConfig, server, syntaxFeatures, polyfill } = userConfig;
   const userConfigHash = await getFileHash(path.join(rootDir, fg.sync(configFile, { cwd: rootDir })[0]));
 
-  await setEnv(rootDir, commandArgs);
   const coreEnvKeys = getCoreEnvKeys();
 
   const routesInfo = await generateRoutesInfo(rootDir, routesConfig);
@@ -188,8 +189,26 @@ async function createService({ rootDir, command, commandArgs }: CreateServiceOpt
     rootDir,
     runtimeDir: RUNTIME_TMP_DIR,
     templateDir: path.join(templateDir, 'exports'),
-    dataLoader: userConfig.dataLoader,
+    dataLoader: Boolean(userConfig.dataLoader),
   });
+
+  if (typeof userConfig.dataLoader === 'object' && userConfig.dataLoader.fetcher) {
+    const {
+      packageName,
+      method,
+    } = userConfig.dataLoader.fetcher;
+
+    generatorAPI.addDataLoaderImport(method ? {
+      source: packageName,
+      alias: {
+        [method]: 'fetcher',
+      },
+      specifier: [method],
+    } : {
+      source: packageName,
+      specifier: '',
+    });
+  }
 
   // render template before webpack compile
   const renderStart = new Date().getTime();